+++ src/gillm/nlp_bridge/__init__.py
@@ -1,5 +1,5 @@
 """NLP translation bridge using nlp2dsl."""
 
-from gllm.nlp_bridge.client import NLPBridgeClient
+from gillm.nlp_bridge.client import NLPBridgeClient
 
 __all__ = ["NLPBridgeClient"]