+++ CHANGELOG.md
@@ -7,13 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2026-06-03
+
+### Docs
+- Update CHANGELOG.md
+- Update README.md
+
+### Test
+- Update tests/test_gillm.py
+
+### Other
+- Update .idea/pyLspTools.xml
+- Update uv.lock
+
 ## [0.1.3] - 2026-06-03
 
 ### Docs
 - Update README.md
 
 ### Test
-- Update tests/test_gllm.py
+- Update tests/test_gillm.py
 
 ## [0.1.2] - 2026-06-03
 
@@ -26,12 +39,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update README.md
 
 ### Test
-- Update tests/test_gllm.py
+- Update tests/test_gillm.py
 
 ### Other
 - Update .env.example
 - Update .idea/.gitignore
-- Update .idea/gllm.iml
+- Update .idea/gillm.iml
 - Update .idea/inspectionProfiles/Project_Default.xml
 - Update .idea/inspectionProfiles/profiles_settings.xml
 - Update .idea/modules.xml